+++ R/check_S2tilinggrid.R
@@ -17,10 +17,10 @@ check_S2tilinggrid <- function(path_S2tilinggrid = NULL){
     zip_S2tilinggrid <- 'Sentinel-2_tiling_grid.zip'
     url_S2tilinggrid <- 'https://sentiwiki.copernicus.eu/__attachments/1692737/S2A_OPER_GIP_TILPAR_MPC__20151209T095117_V20150622T000000_21000101T000000_B00.zip?inst-v=d9cdf66c-96ab-4a2c-8d65-cc56a8f018b2'
     if (curl::has_internet()){
-      download.file(url = url_S2tilinggrid,
+      utils::download.file(url = url_S2tilinggrid,
                     destfile = zip_S2tilinggrid, method = 'auto', quiet = FALSE, mode = "wb")
       tileS2_kml <- file.path(dirname(path_S2tilinggrid),'tileS2_kml')
-      unzip(zipfile = zip_S2tilinggrid,exdir = tileS2_kml)
+      utils::unzip(zipfile = zip_S2tilinggrid,exdir = tileS2_kml)
       path_S2tilinggrid_tmp <- list.files(tileS2_kml,pattern = '.kml', full.names = T)
       file.rename(from = path_S2tilinggrid_tmp, to = path_S2tilinggrid)
       file.remove(zip_S2tilinggrid)