+++ NEWS.md
@@ -4,6 +4,7 @@
 - added input variable keepCRS in function get_s2_raster to preserve native projection provided by aoi_path or bbox
 - modifications to avoid notes when building package
 - adjust crs for aoi and bbox
+- suppress warning when tmp files cannot be suppressed
 
 # preprocS2 v2.2.6